requirements.txt
streamlit>=1.32
altair==4.2.2
numpy
pandas
torch --index-url https://download.pytorch.org/whl/cpu
ultralytics==8.2.103
pillow
